Move CSS scanner helpers to focused partial (#4107)
