+++ changelog.d/unreleased/4107.internal.md
@@ -5,6 +5,7 @@ issues:
 affected:
   - src/CodeIndex/Indexer/Symbols/SymbolExtractor.cs
   - src/CodeIndex/Indexer/Symbols/SymbolExtractor.Css.cs
+  - src/CodeIndex/Indexer/Symbols/SymbolExtractor.CssScanner.cs
   - src/CodeIndex/Indexer/Symbols/SymbolExtractor.JavaScriptTypeScriptStyledFactory.cs
   - docs/large-file-decomposition-plan.md
 ---
